Implement RC Setter for ReturnCode
In order to propagate the multicast compare errors correctly
(with the correct FAPI RC and Hostboot errl), we need a way
for ReturnCode class to be able to reset the internal RC
to somehting other than the default RC without using constructors/
converters. This will allow to maintain a valid plat error log pointer
and have a custom FAPI RC packaged into the ReturnCode structure.

Change-Id: I842cdb2da015a570f3d43cc4e951efe253851823
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87472
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Reviewed-by: Matt K Light <mklight@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/87659
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>