p10_sbe_fastarray: Pull scan region from control stream, work around HW514994
The scan region information is now embedded in the control data; no more need
for the calling code to figure out what region/type combination to use. The
information is decoded/expanded inside the procedure, and a workaround for
HW514994 applied on top if necessary (need to enable BIST for CL2 but clock
MMA if MMA is selected).

Plus a few bits of refactoring - extracting constants etc.

Change-Id: I5018dbd84e56b305d8dd9131c5cedc8107d8bee7
Original-Change-Id: I93bce5b0f88350abd68b25318b3436426897c016
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89006
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Christopher M Riedl <cmr@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102034
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>