vbu/contained: Use correct hangpulse regs, improve chkpt, fix L3 LCO ID
Write the correct registers to disable the hang pulses in RUNN IPLs.

In cache-contained RUNN mode each active L3 needs to be configured
identically. Clear the L3 MODE_REG1 when restoring the L3 config during
the contained re-IPL.

Read a tag from an environment variable to use as the base name when
generating a checkpoint. This ensures the simulation environment has
some control over creating unique checkpoint filenames.

Also refactor the cache-contained loop to start RUNN via scom to the
OPCG_REG0:
instead of ...

	for each eq:
		do hld dly setup
		take checkpoint
		do RUNN start

... do ...

	for each eq:
		do hld dly setup

	for each eq:
		take checkpoint
		do RUNN start

Now every EQ clock-controller is primed to start RUNN when loading the
first EQ checkpoint. The HLD DLY (per core offset) is enabled via a bit
in the OPCG_REG0 so the RUNN start cannot be a multicast op in
cache-contained mode when supporting some EQs running with the HLD DLY
and others running without.

Change-Id: If468e7ad04d1de5a42484be170faad3e5f02473a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89472
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108846
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>