vbu/contained: Use PUTFAC instead of STKFAC to apply MCD parity
The MCD engd does not correctly specify which parity bits to set for
SPYs. Until a model w/ a fix is released, set the proper parity bit via
direct simulator access. Previously, this access was a 'stick'; however,
this is problematic since it prevents scanning of the n1_fure ring. Use
a 'put' instead.

Change-Id: Id0fcf71aae48b6c23cc0c1d30e0c3569c07b4618
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89533
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108847
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>
Tested-by: RAJA DAS <rajadas2@in.ibm.com>