Adjust idec algorithm for Explorer B.0 format
The update to B.0 doesn't allow for a change to the register that
previously contained the MAJOR_REVISION value.  Because of that
we are revising the logic that determines the EC level to be a
lookup based on a single number instead.

Ignore the MAJOR_RELEASE field from CHIP_INFO entirely, it will
always =1. Instead we will rely only on the EFUSE bits currently
designated for MINOR_RELEASE. Since we lost the ability to have
a common M.m we will just use a single incrementing value to
represent the ordinal revision number, regardless of major/minor,
i.e. the 1st revision is 0, the 2nd is 1, the 10th is 9, etc.
So that gives us these values for EFUSE_IMAGE_OUT_3[13:10]
  A.0 = 0 = 0x10 in Host Firmware
  A.1 = 1 = 0x11 in Host Firmware
  B.0 = 2 = 0x20 in Host Firmware
  B.1 = 3 = 0x21 in Host Firmware (if it were to happen)

Change-Id: I878491e8e004c231ce4dcd7abe887a947c7f3dc3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89764
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/89793
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>