p10_sbe_fastarray: Flush array pipelines before starting dump
Inside some arrays there is very short pipelining behind the muxes that
switch control over to the ABIST engines. If a write is stuck in these
pipelines when we stop the cores, it gets executed during the first
ABIST cycles of the dump, corrupting the array contents.

Fix this by running a few cycles without ARY clocks before we start
dumping; that way the writes will flush out of the pipelines without
causing actual write pulses to the array cells.

Change-Id: Ia5cffee6c65f056bf724487837e3630bef15beea
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/90291
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher M Riedl <cmr@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101697
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>