Use full UART TX buffer (shave 5s off booting)
This patch makes us use the full TX buffer of the UART rather than
waiting for each character to go through. This practically means we
get a *massive* IPL speed up.

Measured on my Raptor Blackbird with P9N DD2.21, from "Welcome to SBE"
to "Booting Hostboot" (timed on a stopwatch):
Before (with UART output): 15 seconds
With this patch (with UART output): 10 seconds.

A full 33% reduction in boot time for SBE is nothing to sneeze at!

Change-Id: Iac50d4b820f3b998e55e5f6e0bff60b067c1df45
Signed-off-by: Stewart Smith <stewart@flamingspork.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/90587
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Andrew R Jeffery <andrewrj@au1.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>