PLAT target.C: Make multicast decomposition dependent on actual target type
Multicast target decomposition in getChildren() currently enters the relevant
code paths (core MC vs. non-core MC) based on the incoming target's template
type. For composite targets this is not decidable and the code will always
enter the non-core path for a comp target even if it really is a core MC
target - which doesn't work.

Fix by using the actual target type in the plat target handle for deciding
instead. Also move the "unicast target to its own type" case to the top.

Change-Id: Iec2593d6a730842e8416d9182765530759dd2180
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/92892
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>