Added new attribute ATTR_MFG_FLAGS to supplant ATTR_MNFG_FLAGS
The new attribute, ATTR_MFG_FLAGS, is a 32 bit array of size 4 that
can hold ALL the manufacturing flags, whereas the old attribute,
ATTR_MNFG_FLAGS, is a 64 bit integer that can only hold a subset
of the manufacturing flags.

Also added all the flags (as enums) as found in file:
/fips950/src/mnfg/mfg/mnfgPolicyFlags.H
which match section 6.1.9 of the document:
Hypervisor Interface Data Specifications v10.5o-1.pdf.

Change-Id: I8fd71aa7778bbaf7ea6499939960f5c979c21118
RTC: 249179
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94460
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94467
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>