Fix in p10_ipl_customize for "invalid buffer" fail
This commit fixes the "invalid buffer" and occassional seg fault
encountered when running p10_ipl_customize_wrap.exe.

The root cause is a buffer overrun on ringBuf1 when calling
_rs4_compress during a memset() operation.

The commit also cleans up the naming of the ring buffer argument
for ring buffer 1 to make it more clear that the arg is used to
describe the buffer more so than the Vpd ring.

Key_Cronus_Test=XIP_REGRESS_SBE_QME

Change-Id: Ib2dc0902aaa9e0266ee30efb46595eee0154c552
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94615
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Reviewed-by: Chirag Sharma <chirshar@in.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/94645
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>