FAPI2: Add target method to safely construct reduced-type targets
Add a new method reduceType() to fapi2::Target which converts the
target into a new target with a narrower template type if and only
if the platform target's real type would still fit the new template
type.

One basic principle of FAPI2 is that a procedure should stay
oblivious of the actual plat target type, but if the procedure must
accept a wide range of possible targets (such as putRing) it may be
impossible to write completely oblivious code, let alone generic
platform support that ensures correct operation of getParent and
getChildren across all possible combinations of parent and child
types.

One solution is to allow different code paths in procedures based on
other factors (such as the ring ID), and narrowing down the allowed
set of target types based on the code path. That's where reduceType()
comes into play - it's a way for the procedure to assert a given
set of target types going forward, giving it the opportunity to fail
if that assertion is violated.

Change-Id: Ia535905aa8b418d692cb1288d43272533acfec6c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96011
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt K Light <mklight@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>
Reviewed-by: Jennifer A Stofer <stofer@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96064