ringUtils: use Target::reduceType to simplify target mangling
The broad composite type of TARGET_TYPE_ALL_MC for scan targets
makes it very hard (if not impossible) for getParent and getChildren
to do the right thing in all possible cases.

Based on the code path we're taking the set of valid target types
will shrink - model this by using the new reduceType method to reduce
ambiguity for getParent and getChildren calls (and weed out invalid
targets in the process).

This means that some code has to be moved around, but the end result
should be more readable too since code ends up more local to the path
it serves.

Temporarily add Cronus implementation of Target::reduceType() here until
eCMD is updated with its own implementation.

Change-Id: I5422324de291e2d78be89ee95e51fd82b0d85609
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96347
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Prem Shanker Jha <premjha2@in.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/97938
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>