Rename customrc to projectrc, add back option for real customrc
The idea of customrc is for devs to add their own custom env vars,
so it should not be under source control. Rename the current customrc
which is under source control into projectrc and extend env.bash to
source both files, with customrc sourced after projectrc so it can
override projectrc settings.

Change-Id: Iaf119be938e223c242bc32aa513ea6e4e63cca61
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/96898
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>