istep3: Do not scan eq_clkadj rings in contained IPL
The EQ CLKADJ region clocks remain running during a contained IPL and
are only scanned once during the pre-contained IPL. Unlike other
regions, the initial IPL is responsible for scanning and configuring
this region and the setup is left intact during the contained IPL.

Until now there was no content for the `eq_clkadj_*` rings and so these
rings were always skipped during IPL. Recent VPD testing with actual #G
and #R content exposed a bug wherein the contained IPL attempted to scan
rings in this region. This is impossible since clocks are still running.

Unlike the scan0 procedures, the istep3 initf procedures do not take a
region bitvector and therefore must now be aware if they are running in
a contained IPL context or not.

Change-Id: I3393b37cd330d0a14883e63e3d1320c1e136241a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99216
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Jenny Huynh <jhuynh@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99220
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>