Work around PCB clock alignment problems
When switching the nest mesh from 1:1 to 2:1, the distributed
nature of the clock mux responsible causes parts of the nest mesh
to receive an additional fast cycle before its mux switches to
the slow clock. This causes random parts of the PCB to fall
out of alignment with the rest and become unreliable.

Work around this by quiescing the nest mesh before switching the
mux. There is no architected way to glitchlessly quiesce the nest
once started, but since the nest DPLL is not yet locked we can
take it out of bypass (which glitchlessly switches to the
disabled DPLL output), then reset the nest clock muxes (which is
not glitchless but we're already quiesced), then put the DPLL
back into bypass (which is also not glitchless but the muxes are
now blocked).

Also make sure that align pulses in the PERV chiplet are disabled
before we switch to asynchronous.

Change-Id: I68bdcfb4219fffc261920b997c9bd80bf75bf080
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100654
Dev-Ready: Joseph J McGill <jmcgill@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Justin D Ginn <justin.ginn@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100656
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>