vbu/contained: Fix chip-contained istep4 IPL multicast group
Commit 85c06e7b8 ("p10_sbe_select_ex: Only needed EQs in MC group 6 for
istep 4") changes how EQ chiplets are included in the ISTEP4 EQ_GOOD
multicast group. The new setup uses the pgood data in the CPLT_CTRL2
register which is lost during the chip-contained istep3 re-IPL steps and
restored later. For a chip-contained IPL this setup must use the ATTR_PG
values isntead of the register data so use the new SELECT_EX_MC_GROUPS.

Cache-contained IPLs were not affected since the EQ chiplets are not
re-initialized and the pgood data in the CPLT_CTRL2 is retained.

Change-Id: I0a5a1a71fbeb6c3f4f1d702395533102bc81fb03
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101809
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108872
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>