Set l3_base on all processors to be 4 GB - 256 MB
Simics is not correctly setting l3_base on processor model themselves
and this value must match ATTR_HOSTBOOT_HRMOR_OFFFSET which the SBE
is responsible for setting. Until simics dynamically sets this value
based on HW setting the SBE code makes this value needs to be set in all
of SBE's startup.simics scripts.

Change-Id: Ie5b1406414e3baeab3659c1d41a6460847d3a744
RTC: 207396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104807
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>