vbu/contained: Fix incorrect regions passed to scan0 module
The scan0 module expects regions to be passed as a right aligned
uint16_t. When calling the scan0 module, the region bits were wrongly
extracted from the contained regions constant. This resulted in the scan0
module not actually scan0'ing most of the regions. On DD1 HW there
appear to be no consequences since the sbe_arrayinit function will scan0
all the regions of each chiplet again anyways. During DD2 sim a
checkstop occurs during arrayinit which causes the IPL to fail. The
checkstop appears to be caused by a FIRMASK which is not scan0'd to all
1s.

Pass the correct regions value to the scan0 module for DD2 before
calling sbe_arrayinit.

Change-Id: Ifaee563a897091fd5495eff5a722159e2e77c4fe
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108094
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamin Gass <bgass@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108882
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>