p10_sbe_initf -- skip eq_inex scan for DD1
eq_inex contains functional latches on DD2 that will require scan
initialization, especially for cache and chip contained

The DD1 ring definition is retiming latches only, and is pruned from
the finalized scandef as a result

108115 added a putring call to process eq_inex, with the expectation
that no data will be scanned for DD1.  The Cronus platform emits
warning messages related to the scandef even though there is no
cotnent to scan -- this change simply skips the putring call for
DD1 to avoid the spurious Cronus platform messages.

Change-Id: I36dc71db0d924cbb10c605a3af412940d8fcc0a7
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108222
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamin Gass <bgass@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108224
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>