Fix for SW511601
In order to RAM the SPRs/FPRs/VSRs, the GPR0 and GPR1 are used
to transfer data to/from those registers and the SPRD.  When the
SPRs/FPRs/VSRs are RAMed using the allow_mult=true option, it
will leave the GPR0 and possibly the GPR1 in an invalid state
until such time that the ram_cleanup procedure is called, at
which time the GPR0 and GPR1 will be restored back to a valid
state from their backup values.  If get_reg is called for GPR0
or GPR1 with allow_mult=true before the ram cleanup procedure
is called, then the get_reg procedure will return this invalid
value.  The fix is to force the ram_cleanup/ram_setup procedure
to be called before RAMing the GPR0/GPR1 when these registers
are determined to be invalid.  This will return the GPR0/GPR1
to a valid state.  Another way around this defect is to avoid
grouping together the RAMing of SPRs/FPRs/VSRs and GPRs with
a single pair of ram_setup/ram_cleanup calls.

Change-Id: Id1550266f81e792a4891a00973311ef077706fb7
CQ: SW511601
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108474
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108476
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>