vbu/contained: Fix dump IPL
The dump IPL requires that the core being accessed be configured for PBA
accesses. Previously, the core was initialized in a way that was
conducive to this. Because inits have changed, that all cores are
disconnected from the fabric on re-re-IPL, we need to ensure that the
core being configured for dumping is properly connected to the fabric.

The original concept, of re-IPLing while maintaining L2/L3 array
content for the purpose of dumping, remains the same.

Change-Id: I78ae3226fc39e5b85234327385c0e0e3973d825c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/108809
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher M Riedl <cmr@ibm.com>
Reviewed-by: SAMUEL D KIRCHHOFF <kirchhoff@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/110090
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>