Align xip sections to nearest 8byte before sigining SBE_FW section
This is to fix an issue with SHA512 algorithm which does an unaligned
access in PPE when there are left over bytes from one xip image.
The left over bytes create issue when we are measuring multiple xip
sections. So we have to make sure that each xip section is individually
aligned to 8byte boundary. That means we have to include some pad
bytes at the end of unaligned xip sections. We have to follow the same
in the verification code which verifies this measured SHA.

Change-Id: Ib585d053dee361c6b984f2aa0072d230059086c0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/111224
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>