TOD: Mask external xstop, use common error mask in tod_init
Additionally mask "external xstop" errors since they're just a
secondary error and should not be re-reported by the TOD.

As it turns out, tod_init was using its own error mask value that
was not in sync with the value used in tp_chiplet_init. Make
the value used by tp_chiplet_init public and use it in tod_init
too.

Change-Id: I9956aca35f8521eff38142fdee55cae6567dd1ae
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114775
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Douglas Holtsinger <douglas.holtsinger@ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114784
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>