Change iv_ram_enable to per-Core tracking variable
The RAMing variable iv_ram_enable is supposed to track whether
the RAMing logic is enabled for RAM mode or not.  Currently it
is setup as a per-thread variable, however the RAM mode enable
logic operates on a per-Core basis.  This change replaces the
per-thread tracking variable with the proper per-Core variable.
It is currently possible to cause a checkstop (due to RAM mode
not enabled when attempting to RAM) with the following sequence
of calls:
1) ram_setup() Core 0 thread 0 -> enables RAM mode for Core 0
2) ram_setup() Core 0 thread 1
... Normal RAMing operations ...
3) ram_cleanup() Core 0 thread 0 -> disables RAM mode for Core 0
4) ram_cleanup() Core 0 thread 1 -> RAMing logic signals checkstop

(The ram_cleanup() HWP itself can perform RAMing before RAM mode
is eventually disabled by the HWP).
The debug of this checkstop is complicated by the presence of
hardware defect HW562739.  This HWP change will cause the above
sequence to not disable RAM mode at step 3, but rather disable
RAM mode at step 4.  This will allow the above sequence to
complete correctly without a checkstop.

Change-Id: I0222275f7666bf70d4e468f4d7b287ff8be233d1
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114843
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/114851
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>