MSR defaulted to ipl reset value in MPIPL reset path
We were jumping off to otprom directly from boot seeprom directly
by setting off the IVPR and IAR, it seems we are getting some external
events which is taking the flow out of OTPROM and into the interrupt
handling code, which is modifying some GPRs in between the OTPROM
execution, which is messing up the flow.
With MSR set to reset value, all external interrupts will be masked.

CQ: SW523128

Change-Id: I3811994b54709ad7ea2d3e4de95249a58a683a63
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/115281
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>