Good PCI Filter For MC Groups
Hostboot needs to update MC groups after istep 10.6 to make sure
the PCI config that it updates in that istep is reflected in MC
groups. This commit introduces a PCI filter that looks at functional
PECs as determined by the platform (NOT pervasive-functional) and
filters appropriate PCI/PEC chiplets out of the corresponding MC groups.

Change-Id: Ibc0263fb2a5751cf92ccba019f8e0f610f2ffc47
CQ: SW513918
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116401
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/116414
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>