RAS Review: p10_common_stopclocks_errors.xml
Added HIGH priority code callouts to all of these errors, also
added TARGET ffdc if there was some laying around to provide some
more context if the error occurs. Targets added as ffdc were also
added as medium priority callouts. It is intentional that these
errors will not perform deconfigure as the path they are used in
is error prone and we don't want to deconfig parts when its not
actually required. Some comments were added to the HWP to express
this intent.

Change-Id: I5501bc6f8c74b89b99c4d8edc7123e5fa4786706
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119284
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: RANGANATHPRASAD G. BRAHMASAMUDRA <prasadbgr@in.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/119294
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>