Unlocking the I2C Engine E after every access and HReset Path
Once SOL is asserted and SBE starts a transaction then lock is auto
granted to SBE. SBE has to release the lock by writing into fast
reset register with bit 0&1 asserted. All the other masters will
experience retry error if they try to access engine E while SBE is
holding the lock.
For case of engine E - If SOL = 1, Only SBE is allowed to
      protected ports/id's
SOL = 1, SBE is not owning the lock still non-SBE masters are not
      allowed on protected id/ports
SOL=1, SBE is not owning the lock, non-SBE masters are allowed to
      unprotected id/ports

Change-Id: I8bc36997bfcea92651182c61a7bfa0b169a03bf0
CQ: SW530486
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/120595
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sandeep Korrapati <sakorrap@in.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>