Using Scratch13 and Scratch16 to support to new Hreset and Mpipl flow
Saving the spi clock divider and round trip delay in Scratch16 for
hreset path, the caller(host/sp) will use scratch reg16 to restore
back scratch reg4, to be used in otprom new HReset path.
Once you go via the IPL reset path, LFR is cleared out, we need to
restore the tpm spi frequency and pau frequency to recalculate tpm
spi divider in the mpipl path. For this we are saving the same into
scratch13 to be reused in hreset path.
One of the procedure which makes use of scratch13 have been updated
to do a read/modify write on the scratch reg13, so that last 2bytes
remains untouched.

CQ: SW533911
Change-Id: I422caec4d76d9f07cec12552ca16d9ec46e6044a
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122252
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122261
Reviewed-by: RAJA DAS <rajadas2@in.ibm.com>