Reduce poll_fw_status Poll Counts
poll_fw_status goes through Hostboot I2C code paths that have extensive
retry logic, so high number of polls is not necessary. Additionally, the
hight poll number makes it so that an OCMB that is having an issue won't
fail the IPL until all of the polls and retries finish, which takes a
significant amount of time.

Change-Id: I897ad57f3741cf7a911c20b0fbfe7821dd8fe7c3
CQ: SW538955
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130688
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN AMMANDA DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130691
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>