p10_sbe_tp_switch_gears: Add delays to 2:1 mux switch sequence
As a tentative fix for IPL problems that look like a reoccurence of
the nest mesh misalignment problems of yore, add a small delay after
each step in the switch sequence.

The theory is that propagation delays cause the DPLL to still output
a clock at the time we reset mux3, causing the mesh clock to glitch.
Adding the delays will prevent that.

CQ: SW544329
CQ: SW547085
Change-Id: Id8536f60ee8322be0dc78db3f0786f1013fd1ed3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130136
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN AMMANDA DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/130139
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Dean Sanner <dsanner@us.ibm.com>