H8: Update calls of python to python2 in migration effort to Red Hat 8
RH7 has /bin/python defined in it's environment whereas RH8 will not.  RH8
will not have a default of /bin/python but will only have /bin/python2 and
/bin/python3 defined.

Considering that for RH7 /bin/python is just a symbolic link to /bin/python2,
making this change to explicitly call python2 will have no detrimental affects
for compiling SBE on RH7 but will position the compiling of the SBE code on
RH8 for success.

Change-Id: Iad5e44caa1debf1111b9520b5e3129044872640f
RTC: 296965
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133094
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Akhilesh S <akhis023@in.ibm.com>
Reviewed-by: Sunil Kumar <skumar8j@in.ibm.com>
Reviewed-by: AMIT J TENDOLKAR <amit.tendolkar@in.ibm.com>